+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 * refactor: spin off `resolveTargetClass()` to separate class
 * perf: proxy warming
 * feat: process proxies of all classes in the inheritance chain
+* feat: throws exception if the paired property does not exist in the other
+  class if the `Map` attribute has `class` set
 
 ## 1.10.0